// frontend/src/app/openapi/model/announcement-response.ts
/*
 * SPDX-FileCopyrightText: Copyright DB InfraGO AG and contributors
 * SPDX-License-Identifier: Apache-2.0
 *
 * Capella Collaboration Manager API
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * Do not edit the class manually.
 + To generate a new version, run `make openapi` in the root directory of this repository.
 */

import { AnnouncementLevel } from './announcement-level';


export interface AnnouncementResponse { 
    level: AnnouncementLevel;
    title: string;
    message: string;
    dismissible: boolean;
    id: number;
}
export namespace AnnouncementResponse {
}